Finish the /bind quoting reworking. (I hope.)
This version takes the (almost) unparsed command string, strips quotation
marks (if the string is a single token due to quoting) and sets that as the
binding, and reverses this if needed when saving the bindings.

This, in testing locally, has been found to be sufficient for multi-command
bindings to be properly useable while allowing for multiple quoted strings
within any given binding (for the /if command, given then or else commands
with parameters).