New command which picks another command from a list according to modifiers (not) pressed.
/modcase <modifiers> <command> [<modifiers> <command>] ... [<command>]

The modifier strings and command strings are the same as for /if.

The command is chosen from those whose modifier strings match the keyboard
state. An effective sort is applied: modifier count (descending) then
parameter position (ascending).

So higher modifier counts win over lower modifier counts; for identical
modifier counts, leftmost wins.

This means that, given

  /modcase ctrl "/echo Foo" alt "/echo Bar" shift,ctrl "/echo Baz"

* "Baz" is output if Ctrl and Shift are pressed;
* else "Foo" is output if Ctrl is pressed;
* else "Bar" is output if Alt is pressed;
* else the default command would be executed, if there were one.

I hope that this is clear enough! ☺