Completely rework the font cache.
The new cache uses one image for each block of 256 code points.
Initially, only block 0 for each font is populated.
More blocks are added as glyphs are needed from them.

If a font is requested but is already cached, the cached copy is returned.
However, there is no reference counting.

Nothing is saved to disk (yet).