remove/simplify a bunch of path-related functions
installPath is only written, by Sys_SetDefaultInstallPath(), and is never read. these can be removed. the only use of Sys_SetDefaultInstallPath() is ``Sys_SetDefaultInstallPath( DEFAULT_BASEDIR );'', which is to be removed. the DEFAULT_BASEDIR macro can actually resolve to calls to Sys_StripAppBundle() and Sys_BinaryPath(), but these calls do not have side-effects. DEFAULT_BASEDIR and Sys_StripAppBundle() are left unused, and thus can also be removed. The only use of Sys_BinaryPath() can be inlined.

Sys_SetDefaultLibPath() is not used, and thus libPath always remains the empty string. these can be removed, and Sys_DefaultLibPath() can be simplified. Sys_DefaultBasePath() is a bogus, droppable declaration.

after the above steps, no functions of the form Sys_SetDefault<which>Path() exist. it makes sense to have a Sys_Default<which>Path() function for each <which> in {Base, Home, Lib, App}. however, Sys_SetDefaultBasePath() is instead called Sys_SetDefaultInstallPath(). the latter is renamed to the former.