Add visvis calculation and cgame trap.
Visvis is similar to vis, but while vis indicates if there is a direct
line of sight between two clusters, visvis indicates that there is an
indirect line of sight (i.e. there is a common point from where there are
lines of sight to both clusters).

The computation of visvis used here is not geometrically exact, but if
it differs from the definition given above, it's always the case that
it may indicate an indirect invisibility where none exists, so it is
still safe to use it for culling. In practise the differences are not
important.