Add Color-Grading support to the GL 3renderer.
The look-up tables have to be loaded into 2D-textures first and then copied to
3D textures as we have no 3D texture loaders and also to avoid an in-game
disk access when the cgame switches the color grading in midgame.

neutral.webp contains a 1:1 color mapping table, it can be modified in Gimp
to derive custom color mappings.

Cgame support is very limited so far, only a /testcgrade <image> command to
load a custom color map.