Cmd: Fix the previous commit
Following ZTM's suggestion VM commands are now registered with the VM
specific function (instead of registered with NULL). To make it work without
changing the gamelogic there are some hacks. TODO: check that the gamelogic
registers all the commands and make cgame forward some commands to the server.

All the hacks are my own creation and not ZTM's