Revert "Rename sel_ldr to nacl_loader". Breaks .deb build requirements.
This reverts commit a748f385f8ee2f674ece6dd35a04e7135678dbd4.

This change causes a download to be started while building a package because
/usr/lib/nacl/nacl_loader doesn't exist. If I rename sel_ldr to nacl_loader,
I break older versions (which some people may legitimately want to run)
unless I add Conflicts headers to the newer nacl packages.

I could add a symlink in THIS package, but that's setting up potential
future breakage.

If this binary is renamed upstream then yes, we need to deal with that.
However, it does not appear to have been.

Therefore, this commit is considered broken.