cmake: Prefer to find the raw lua library instead of a versioned one
The FindLua.cmake function prefers to find the lua library that is
formatted like lua<version>.<os suffix>. This leads to it picking
incompatible headers and libraries on certain distros (notably
Arch). Therefore, if we prefer to find lua.<os suffix> it is more
likely to match with the includes.