add some unreachable builtin on UNREACHABLE() assert macro
The compiler raises some miswarning:

> warning: control reaches end of non-void function [-Wreturn-type]

But we know the code asserts before reaching end of non-void function.
So, instead of adding some useless return after the UNREACHABLE()
assert macro everwhere the compiler raises mistakenly a warning, I added
support for some builtin that tells the compiler this code can't be reached.
If the compiler is too old or is'nt supported, it will raise the useless
warning, which is not a big issue.

The UNREACHABLE() assert macro was renamed to ASSERT_UNREACHABLE() like
other assert macros.