Do additional validation for Keyboard::Key deserialization and cgame->engine calls.
Also allow scancode keys to be constructed anywhere. I had them only
being able to be constructed in engine before, due to the function
depending on an SDL-defined constant, but thinking about it, it seems
quite reasonable for someone to want to construct a scancode key in the
cgame, so I just hard-coded the number.