Require a -uri switch before a command line URI.
The test for the prefix "unv://" as previously implemented is not
reliable, as on Windows (at least) anything beginning with "unv:" will
also go the 'unv' protocol handler. It is important that the
URI-handling code path is used any time that the engine is invoked as a
protocol handler because there is no support for escaping, so arbitrary
arguments can be inserted afterward.