Rename the BUILD_CLIENT preprocessor definition to BUILD_GRAPHICAL_CLIENT
In my opinion the name BUILD_CLIENT holds much potential for confusion,
as one might expect it to be defined for any type of client, while it is
in fact only defined for on of them.

Leaving the BUILD_CLIENT name the same in CMake options as it doesn't
have the same potential for programming errors.