tty: Do not use nonblocking stdin
This is buggy because if the engine is launched from a shell,
stdin will be shared and if anything on the shell (like gdb)
changes the stdin status, then unvanquished will also block.
Instead use select to determine if there is input ready to
be read. I didn't not notice any performance impact from this
change.