Build external_deps DLLs which can be safely called on x86 Windows
By default gcc assumes the stack will be 16 byte aligned on a function
entry. This is incompatible with MSVC code which only does 4 byte
alignment. The problem manifests when SSE instructions are used on the
stack. So decrease the alignment assumption to 4 with the
-mpreferred-stack-boundary=2 flag.

Also explicitly enable SSE2 for x86 Windows builds to match Daemon
flags.