Add support for symlinks within zip packages
Symlink support is controlled by a cvar which is disabled by default.
Use of the feature is discouraged; it is intended for testing Xonotic
assets. A symlink may only link to files in the same zip archive. This
change is designed to have minimal impact on performance for non-symlink
files, while good performance for symlinks is mostly a non-goal.

Thanks to illwieckz for some parts of this commit.