glsl: unify some symbol names
Sometime incident ray is named I, sometime R
while reflection ray is usually named R
and viewDir is sometime named eyeDir
and normal in world space is usually named N
but sometime named N2 because N is already taken
for something else.

It's better to use the same names consistently
accross the source code when it's name the same
thing, and it's better to use full-word variable
names when they are not loop iterators.