Subdirectory support for PAKSERVER checks
The initial implementation requires the PAKSERVER file to exist in every
directory with downloadable content. This commit makes a change to the
download protocol (to be switched on in the next release) which allows
any prefix of the download URL that ends in '/' to be used as the
directory to check.