reliefMapping: revert my mistake introduced in 071c789: the flat texture is not read
the glsl code was written to even not read the flat texture
we may even rewrite the CPU part to not bind the flat fallback
at all when r_normalMapping is disabled.

use same value as flat texture instead (0.5, 0.5, 1) as defined in tr_image.cpp
even if (0, 0, 1) will not be that wrong, just to keep things consistent.

the flat texture is still used when normalmapping is enabled but
material does not provide normal map.