tr_init: make r_normalMapping a latched cvar
the glsl code expects this to be defined before compilation.

not having this a latched cvar means different code is
run when r_normalMapping is switched off but vid_restart
is not yet called, and when r_normalMapping is already off
when vid_restart is called.

the previous code was preventing shader compilation error
to be detected in some scenarii, when r_normalMapping was
switched off then switched on before doing vid_restart.