builtin: import to-be-builtin assets in pkg/daemon_src.dpkdir
The lifegoal of those assets is to be embedded in engine, a bit like glsl shaders are.
Those are the assets required by the engine to run.

Before we find a solution to embed them in engine, shipping them within engine
repository is not only the first step to that objective, it makes easier for
third-party games developing games on Dæmon without forking Unvanquished to start
the engine, by loading the dpkdir from the engine repository.

For example the Xonotic's daemon-glue would be able to do something like that:

    build/daemon -pakpath libs/daemon/pkg -set fs_basepath daemon […]

See https://gitlab.com/xonotic/daemon-glue for Xonotic's Daemon Glue.

See https://github.com/UnvanquishedAssets/unvanquished_src.dpkdir for
the repository those assets come from.