Write condump output to passed file name
When we pass arg to /condump command, it should be taken as file
name where command output should be written to. Now if arg is
passed, it is taken as file name. If no arg is passed, file name
is generated by the engine.

Fixes https://github.com/DaemonEngine/Daemon/issues/304