Use a gamelogic-provided table for player state in netcode
The table which drives the transmission of playerState_t fields was
formerly baked into the engine. This meant that changing anything
required a new engine build. Now, the fields can be freely modified with
only cgame+sgame rebuild (except for a few which the engine uses
directly).

When the cgame or sgame is started, the engine sends it a syscall
requesting the list of player state fields.

Also the netcode is optimized a bit in one spot by removing some
unnecessary copying.