Drop support for cinematic mode and videoMap
The video functionality (either cinematic or shader-based) is currently
broken because the last release of Theora (from 2010!) has an
out-of-bounds indexing bug, which apparently causes heap corruption.

Cinematic mode is unusable by design, because it can't be used at the
same time that the cgame is running.

So remove the video playback code and the dependency on the
apparently-dead Theora library.