Fix deprecation warnings in toolchain-nacl.cmake
This bumps the CMake version to 3.13 and fixes deprecations warnings
that were raised in toolchain-nacl.cmake as well as deletes some code
paths that were for older CMake versions. Not that 3.13 is what's in
Debian stable and oldstable has backports for 3.16.

CMakeForceCompiler is deprecated so instead we set CMAKE_C/CXX_COMPILER
directly. This causes CMake to not set CMAKE_C/CXX_COMPILER_ID (because
it can't determine it automatically) so the CMakeLists.txt had to
support an empty CMAKE_CXX_COMPILER_ID and a special case for PCH
support on NaCl.

Getting a target's location as done in pnacl_translate is deprecated so
instead we compute the paths to the various targets/nexe/pexe directly
from the build dir and the gamemodule dir.

Also remove our local copy of CMakeParseArguments as it is a builtin
function since CMake 3.4.

Fixes #412