Generate GLSL includes with pure CMake
For hex-dumping the GLSL source files into character arrays, use a CMake
script instead of a C++ program. This fixes #122 (include-bin can't be
used with cross-compilation).

Also, put the GLSL generated headers directory in the build directory
instead of the source tree. This makes the build work with the source on
a read-only filesystem.