Run botlib in sgame instead of engine
sgame IPC calls for bot navigation are removed, and one is added for
navmesh debug drawing commands (which can only be used if the server is
running as part of a client).

Bot code will live in the Unvanquished repo from now on, but for this
commit it is kept here to have a nice diff for the required changes.