Make Math::Clamp work on NaNs
It will return the lower limit of the range if given NaN.

Also remove the guarantee that if given 'min' higher than 'max', it will
return 'min'. Instead min > max is illegal.

Also fix the places where Clamp might be called with max > min so
that cannot happen. (I didn't find any of these in gamelogic.)