Remove 'thread_local' in gamelogic (fix Mac dlclose)
On MacOS, dlclose has no effect if the shared library contains a
thread_local variable. Get rid of an instance of thread_local
in gamelogic DLLs to fix this. It is important for gamelogic
DLLs to be unloaded and reloaded so that memory is freed and
variables reinitialized.

There is only one pipe in gamelogic, so thread_local is pointless there
because it's inherently thread-unsafe to read from a singleton pipe.