Resolve master server addresses asynchronously
In the server, resolve the address of master servers in a separate
thread, so as to never block on a slow DNS response.

Each master server now has a separate heartbeat timer, instead of all
heartbeats being sent at once. However IPv4 and IPv6 heartbeats are sent
at the same time for a given server.

Client DNS queries are unaffected.