Disable getinfo when server is not running
When the server is not running, we tell the master servers that we are
dead. So it is annoying that we would need to look up their DNS in
response to this packet. So just ignore getinfo packets.

From a player perspective a server which has no running sgame might as
well be ofline.