Delay GLSL shader build until first map load.
This makes possible for the user to quickly reach the game main menu
without building all GLSL shaders and set a low graphics preset before
joining a game and loading a map.

Doing so prevents building unwanted or unsupported GLSL shaders on slow
and/or old hardware and drastically reduce first startup time.