external_deps/build.sh: add --list-platforms
It makes possible to do something like this:

./build.sh --list-platforms | grep -v macos \
    | xargs -I{} ./build.sh {} required
./build.sh --list-platforms | grep macos \
    | ssh macos xargs -I{} ./build.sh {} required
./build.sh --list-platforms \
    | xargs -I{} ./build.sh {} install package

In this example the current system is running Linux
and a macOS host is accessible via SSH with the
macos host name. Everything but macOS package is built
on Linux, macOS package is is build on macOS host,
and every package is installed and packaged on Linux.