external_deps: use LLVM instead of Microsoft lib.exe
This was mostly done to avoid having a step that must run on Windows.
But also there is a change in soft_oal.dll from the latest OpenAL that
makes the gendef output always have the name "soft_oal.dll" (instead
of matching the input file name OpenAL32.dll), so we have to override
this name. llvm-dlltool has the -D switch for this, while lib.exe
seems to lack an equivalent.