cl_input: Fix typing +command in console for repeat key press.
If someone wants to type a +command like +speed or +attack in console it
doesn't work (even if they add a random arg with it) because this check
will always pass and it returns. However, since +commands in console
never have a key asscoiated with them, check whether the key is valid
first.