Start VM process with posix_spawn instead of vfork
The motivation is vfork being deprecated on Mac. But also it
simplifies the error handling - we no longer need to create
a socket pair to report errors that occur betweeing forking
and execing.

Glibc since at least 2.27 uses a vfork(-ish) implementation of
spawn, so it won't lose performance.

It could potentially break something related to NaCl gdbserver
debugging, but no one has managed to use that in many years
anyway.