Enable variable shadowing warnings
-Wshadow=local-compatible seems better than -Wshadow=local, but I chose
the latter for its equivalency with the MSVC warning.

The code changes are to fix conflicts of a lambda argument and variables
in the enclosing function. I think the warning should not complain about
this but there is not fine-grained enough control of it. I deleted some
offending code in Vector.h since we are deprecating that anyway.