Complete cvars up to a namespace
'_' and '.' are considered to be namespace separators. When completing a
cvar name, a group cvars sharing a common namespace will be collapsed to
just the namespace separator, if the group has at least 2 cvars but does
not include all the cvars.

For example, "g_bot_" or "audio." are cvar namespaces.