Introduce unit testing with Googletest
Add tests for Cmd::FoldNamespace().

Tests can be run by enabling the BUILD_TESTS CMake option and running
the test-dummyapp, test-server, test-ttyclient, or test-client binary.

Googletest is included as an (optional) submodule.

Tests are built directly into the engine to work around the codebase's
lack of modularity. For each engine flavor, a test Application is
defined which inherits from the non-test variant and customizes some
behavior to be more suitable for tests. The point of this is to take
advantage of the knowledge of how to initialize the global state in
Application::Initialize(), so that tests do not have to prepare the
global state themselves.