Fix loading unused permutations from shader cache
The renderer attempted to load the cached compiled version of a
shader permutation *before* checking whether it actually wants to build
that permutation or not. I had a file in the cache matching the filename
for an unused permutation, which was built with a different graphics
driver version. This led to an "Invalidating shader binary cache"
message on every load (which would have been very bad and broken caching
generally if the shader in question had not been one of the last ones to
be compiled).

Fix this by checking whether a permutation is used before trying to load
a cached version. Cleans up a bunch of reduntant checks for
unusedPermutation.