Don't assume NaN and Inf cannot exist
This change applies to both Daemon and Unvanquished.

Prevents issues such as https://github.com/Unvanquished/Unvanquished/issues/2299,
because now "nan > x" isn't true half of the time like it was before.
Instead it is always false so the collision condition will never be
entered.

Fixes https://github.com/Unvanquished/Unvanquished/issues/2311, because
most of the nonsensical results tend to have to do with fast-math doing
garbage with nan and infinities, which is now disabled. The rest of
fast-math is supposedly fine, or at least acceptable in our case.