Regenerate shader text when renderer restarts
GLSL_InitGPUShaders was mistakenly grouped with the code that only runs
if the GL context needs to be recreated. This meant the shader text is
only regenerated if you change something about the window or invoke
/vid_restart. But it should run every time the renderer restarts
(e.g. on map change). It is necessary to regenerate the text because
it may be affected by (latch) cvars.

To see the bug, change the value of r_dynamicLight and then change map.
The change will not take effect (despite the cvar being unlatched) and
in some cases the engine will crash (if dynamic light shaders were
never built but are used).

To be clear, this does NOT cause any additional shaders to be built
as long as no cvars were changed. We ALREADY compile or load from cache
each shader on every renderer restart. The performance impact on my
machine is less than 2 ms.