Remove Sleep() call from PDCurses
I've known about problems with low FPS from the Windows curses console for
years but hadn't gotten around to fixing it. Sleep() uses a
low-resolution timer so attempting to sleep for 1 ms may lead to
sleeping for up to 15 ms.

Today it was causing some sort of weird problem in daemonded where the
server was hanging for 30+ seconds so I was motivated to fix it.

I had been intending to replace it with a higher-resolution timer but
actually there is no legitimate to reason to sleep at all so just don't
sleep.