Platform: do not require to duplicate dll and exe extensions and path separator defines
Windows is very likely the only one requiring something different,
and if exists, the other systems not doing things the Unix way
may be even more rare or have least interest than systems like
Android, BSD variants, and others.

On that part of the code, it's now only required to add an ifdef
and a define (two lines) for setting a new platform if it's WIN32
or UNIX compatible.