tr_bsp: rework precomputed light, light map, deluxe map enablement
The render is now as expected not only when r_precomputedLighting
is enabled but also when it is disabled.

Disabling lightmapping while keeping deluxemapping enabled
now still loads the deluxemaps to use them with gridlighting.

A variable named tr.worldLightMapping is used to know when
light mapping is enabled instead of redoing all the tests
with potential mistakes. The is similar to the already
existing tr.worldDeluxeMapping variable.