Fix "bad portal" handling (permanently broke renderer)
Thanks to illwieckz for finding the cause of the bug.

illwieckz: "This may happen if, for example, a portal is seen while being outside the map.

There can be other broken ways to reproduce the bug.

The bug can be reproduced with Spacetracks map 1.1.3
from Unvanquished 0.54 and: /setviewpos 378 -1906 625 80 1

This would break the render, even loading another map
would not fix the render.

To fix the render, one would have to do: vid_restart
or move to a place where a portal is properly rendered
like this: /setviewpos 27 -1838 668 147 13"