Simplify mouse mode handling
The mouse_mode variable now always represents the actual state of
the program. Also, IN_SetMouseMode will never set the program state
to something not requested by the caller except in the case that
nograb is on and Deltas is requested.

Add a log cvar logs.level.client.mouse to log state
changes.