Fix wrong startsolid/allsolid trace results
There were early returns in the case where a trace fraction had
been detected to be 0. However, sometimes the fraction is 0 against some
brush, but not considered to be intersecting it. (This often happens
because the movement traces try to stop where the distance would be
exactly zero.) In that case, we need to keep checking for other brushes
in case one of them is intersecting at the starting point of the trace.

For example on plat23 this trace produced a false negative for
startsolid:
	start (-243.4 1708.8 27.1)
	end (-235.4 1700.8 27.1)
	mins (-23 -23 -22)
	maxs (23 23 14)