Set Windows DPI awareness in code
Set it for the graphical client only as pdcurses does not know how to
scale.

We have a DPI awareness manifest, but it is ignored by the MinGW
toolchain. I know how to fix it but it is pretty messy when combined
with the fact that I want to set it for the graphical client only, so it
seems better to do this in code.

Fixes https://github.com/Unvanquished/Unvanquished/issues/2243.