Fix manifests on MinGW and remove DPI manifest
Adding manifests as a source in cmake does nothing with mingw
toolchains. You have to add the manifest (there may only be one) via a
.rc file.

Remove DPI manifest as this is now done in code, and we don't want to
say that the curses-based applications are DPI-aware. Also the manifest
specified the per-monitor variant which is not right for Daemon as it
does not respond to an event of moving the window to a different
monitor.

Also reorganize the file structure of Windows resources.