renderer: introduce r_lightMpde, implement vertex lighting, revamp the lighting selection, fix bugs
- Introduce r_lighting cvar with values:

0: fullbright
1: vertex lighting
2: grid lighting
3: light mapping

The default value is 3 (light mapping), users may also use 1 (vertex
lighting).

All other values use the light mapping code but with different inputs,
so none of the others are expected to be faster than light mapping.

Values 0 (fullbright) and 2 (grid lighting) may give unfair advantages
without satisfying a need for performance, and are considered cheats.

The only thing grid lighting or vertex lighting can do is to save GPU
memory by not uploading light maps, but the code itself is not faster,
and is even a bit slower with grid lighting because there is more
computation to do.

- Implement fullbright lighting as a debug option.
- Implement vertex lighting for both in-game usage when needed and as a debug option.
- Use vertex lighting for bsp surfaces without lightmap.
- Deeply rewrite the code to make it matching more the description of the need,
  make it less implicit, and track down bugs.
- Limit usage of Half Lambert Lighting and Wrap Around Lighting for models.
- Wrap Around Lighting uses the value set in material.
- Properly enable Rim Lighting.
- Define more GLSL macro conflicts to help the engine skip unused combinations.
- Recognize light-styles lightmaps and process all light-styles
  lightmaps with Render_generic to avoid dynamic lights.
- Properly set implicitLightmap on textures without materials.
- Do not render lightmaps on collapsed stages without lightmaps.
- Make fullbright and grid lighting for world rendering cheats.

The engine may still use fullbright rendering or light grid rendering
when needed (models use light grid) or as a fallback.

renderer: correct light style implementation