Remove unused git revision info
There was only one reference to REVISION, in qcommon.h, but it was never
used because the relevant header was not included.

It's not implemented reliably anyway. The revision detection only reruns
at random times (when cmake stuff is modified), so it is only reliable
for a freshly created build dir, like for a release. But for releases we
just hard-code the version with update-version-number.sh. The revision
could be made always correct by forcing the revision header to be
re-created on every build but that's ugly since it slows down
incremental builds.