cmake: detect arm compiler supporting hard float but ignore the target version
In more recent distribution like Debian 12 Bookworm
or Ubuntu 23.10 Mantic Minautor, it has been observed
that arm-linux-gnueabihf-gcc sets:

 #define __ARM_ARCH 6
 #define __ARM_ARCH_6__ 1
 #define __ARM_PCS_VFP 1

Previously it was setting __ARM_ARCH to 7.

So it looks like the default target is armhf with armv6 as a baseline
instead of armhf with armv7 as a baseline. Even if we will build for
armv7 with neon, we need to detect it's an armhf compiler and we may
just ignore the default arm version. What we need to know is if
the target is using hard float ABI, hence __ARM_PCS_VFP.

In fact it even exists some armv5 with hard float, so we better not
check the default target, all we want is to detect hard float, and
from that we set our default target anyway.