Better detect incompatible versions of game VMs
The previous mechanism was that one should update some constants
CGAME_API_VERSION and GAME_API_VERSION, but no one ever did it. The
counters only got to 3 in 10 years of development.

Instead tie gamelogic ABI versioning to "major" Unvanquished releases. The
assumption is that in each major release there will always be a few
compatibility-breaking changes, while a point release definitely should
not have any such changes or else the previous gamelogic can't run.

This will help users get an informative error message when they try to
replay a demo or run a mod that is not compatible with the current
Daemon.

Additionally there is a constant defined for whether the gamelogic ABI
contains breaking changes intended for an upcoming release. This should
be set on the next-release branch, for-X.Y.Z/sync. The detection of
discrepancies in the compat break bit will help developers keep things
straight when they are switching between the upcoming release branch and
other branches, forget to rebuild stuff, set the wrong vm type flags etc.
This can also be useful when bisecting the engine as you could try to
avoid commits from the compat-breaking branch to avoid having to rebuild
gamelogic every time.

Updates to Unvanquished's update-version-number.sh will help to
automatically update the ABI version and compat break bit.