cmake: make -march flag unavailability an error
If we don't report that the configured -march is an error
we then get random error elsewhere, like a report that C++ 14
is not available, not because C++ 14 flag raises an error,
but because the -march one does.

If the -march flag is unavailable, then we cannot use
the USE_CPU_GENERIC_ARCHITECTURE option so something
is broken and the user should know.